+++ Kernel/Output/HTML/TicketMenu/Move.pm
@@ -122,8 +122,9 @@ sub Run {
         );
         $MoveQueues{0} = '- ' . $LayoutObject->{LanguageObject}->Translate('Move') . ' -';
         $Param{MoveQueuesStrg} = $LayoutObject->AgentQueueListOption(
-            Name => 'DestQueueID',
-            Data => \%MoveQueues,
+            Name  => 'DestQueueID',
+            Data  => \%MoveQueues,
+            Class => 'Modernize',
         );
     }
     else {